Installation modified
Add a check for the log before installation (installation is still disabled by default because we are not ready with a new and fresh ISO). Also the log is recreated before ending the process of formatting if there were some errors: the aim of this program is to find faulted disks, not to recover, for this reason greater details about the faulted disks are added in log instad of a long tail of badblocks errors. Changed also parameter for smartctl: from -i to -ax do add more details